+++ moit-api/src/main/kotlin/com/mashup/moit/controller/sample/SampleNotificationController.kt
@@ -2,7 +2,7 @@ package com.mashup.moit.controller.sample
 
 import com.mashup.moit.common.MoitApiResponse
 import com.mashup.moit.infra.fcm.FCMNotificationService
-import com.mashup.moit.controller.sample.dto.SampleNotificationRequest
+import com.mashup.moit.infra.fcm.SampleNotificationRequest
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
