+++ moit-api/src/main/kotlin/com/mashup/moit/controller/sample/dto/SampleDto.kt
@@ -1,6 +1,5 @@
 package com.mashup.moit.controller.sample.dto
 
-import com.mashup.moit.domain.moit.NotificationRemindOption
 import com.mashup.moit.domain.sample.Sample
 
 data class SampleCreateRequest(
@@ -15,20 +14,6 @@ data class SampleResponse(
         fun of(sample: Sample) = SampleResponse(
             id = sample.id,
             name = sample.name,
-        )
-    }
-}
-
-data class SampleNotificationRequest(
-    val studyId: Long,
-    val studyName: String,
-    val remainMinutes: NotificationRemindOption,
-) {
-    companion object {
-        fun sample() = SampleNotificationRequest(
-            studyId = 1L,
-            studyName = "전자군단",
-            remainMinutes = NotificationRemindOption.BEFORE_10_MINUTE
         )
     }
 }
