+++ moit-api/src/main/kotlin/com/mashup/moit/controller/study/dto/StudyDto.kt
@@ -93,28 +93,3 @@ data class StudyUserAttendanceStatusResponse(
         )
     }
 }
-
-@Schema(description = "Study 출석 시작 notification")
-data class StudyAttendanceStartNotification(
-    @Schema(description = "moitId")
-    val moitId: Long,
-    @Schema(description = "notification title")
-    val title: String,
-    @Schema(description = "notification body")
-    val body: String,
-) {
-    companion object {
-        private const val TITLE_TEMPLATE = "출석 체크가 시작되었어요!"
-
-        // todo edit push message depends on notification level 
-        fun of(moit: Moit, study: Study) = StudyAttendanceStartNotification(
-            moitId = moit.id,
-            title = TITLE_TEMPLATE,
-            body = bodyTemplate(moit, study)
-        )
-
-        private fun bodyTemplate(moit: Moit, study: Study): String {
-            return "${moit.name} ${study.order}회차 스터디가 시작되었어요"
-        }
-    }
-}
